Blogs: more work on the update logic
- to separate the code a bit based on it's purpose, I've added a "BlogPostsWebClient" for making the requests to the interwebz (fetching and parsing the RSS feeds)
- the "BlogPostsService" class now has a single "UpdateBlogPosts" method that takes care of everything necessary for updating the cache / files on disk
- the "BlogDatabaseItem" class now has a numeric ID (used as primary key in the database), which we now can use when indexing in Examine. This is better than an incremental counter, since we can now index blog posts on the fly.
